PID-30 #in-progress #comment "BSW refactoring & PI tuning"

+++ MechanumWheelControl/Decision/Decision_State.h
@@ -41,6 +41,6 @@
 /*Global Function Prototype*/
 /***********************************************************************/
 
-uint32 get_state(void);
+uint8 get_state(void);
 
 #endif /* DECISION_STATE_H_ */
